Oops, forgot `get_nearest`. Moved it to utils, too
